Initial commit of ContingentLinearContentProvider.
The ContingentLinearContentProvider makes investment and pricing
decisions (in all but the first step), based on a set of condition
thresholds and what is observed in the environment.  These thresholds
are encoded in the individual's genome.  Each is a separate binary
condition, and may therefore be set to correspond to a binary digit
that, taken together with the evaluation of other conditions, determines
a market condition index.