Various changes to remove a large number of compiler warnings.
This was mostly just specifying unknown generics rather than leaving
these generics unstated.  The only substantive changes were in the
NeutralityModel class itself, where an awkward method of creating cloned
lists to avoid ConcurrentModificationExceptions was replaced with the
more elegant solution of using iterators explicitly and using the
it.remove() function, which is designed to handle this case.  :)