Fix for NetworkOperator bundling.
Previously, network operators could offer either only bundled video
content or stand alone content, depending on whether or not bundling was
enabled at the model parameter level.  

While this worked fine in the single network operator scenario, it puts
network operators at a severe disadvantage when there's more than one
network operator in the system, because they can only sell their content
on their own platform, as compared with third party content providers,
who can sell their content on all platforms.  This means that they could
only receive 1/n as much revenue for the same investment as third party
providers, where N is the number of network operators in the simulation.

On the other hand, allowing network providers to offer both bundled
content and unbundled content on their own network simultaneously puts
them at an advantage, because their quantity from each is considered
separately, rather than having the utility scale with the square root of
quantity.  Essentially, they get x^0.5 + y^0.5, rather than (x+y)^0.5.

To be even-handed about this (and we should try to be as even-handed as
possible), there should be only one option there.  Therefore, the
provider shouldn’t be able to offer a bundle and independent content
simultaneously, but only on the provider’s own network.  This does seem
to be consistent with the real world.

One potential issue though…  what if the unbundled combination is
cheaper?  That is, the provider’s unbundled network price plus its
unbundled content price are less than its bundled price?  At least this
is a learning issue more than a structural fairness one…  It can always
lower the price of its bundle!  And charging more for the bundle would
be giving an advantage to its competitors!  So… we can probably rely on
the learning system to prevent this, at least for non-trivial
differences.  Plus, again, network providers do have this choice in the
real world…